fixed the integer size for outb

+++ src/io.h
@@ -6,7 +6,7 @@
  * @param port the I/O port to send the data to
  * @param data The data to send to the I/O port
  */
-void outb(uint16_t port, uint16_t data);
+void outb(uint16_t port, uint8_t data);
 
 /**
  * Read a byte from an I/O port
